Take a loan of up to 60% loan against total portfolio value

+++ lib/models/loan_model.dart
@@ -1,6 +1,6 @@
 class LoanModel {
   double? amount;
-  int? period;
+  double? period;
 
   LoanModel({this.amount, this.period});
 }
